skills/image-processor-skill/pyproject.toml
[project]
name = "image-processor-skill"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["Pillow"]
